moved all weather package tests into single file
